test(data-loader): in-source
